# api/openapi.yaml
openapi: 3.0.3
info:
  title: Open API Kundenbeziehung - Swiss National Standard
  description: |
    Schweizer Standard für den branchenübergreifenden Austausch von Kundendaten zur Etablierung einer selbstbestimmten digitalen Kundenbeziehung.
    
    Diese API implementiert das vollständige Framework der Open API Kundenbeziehung mit dem universellen 10-Stufen-Referenzprozess und ermöglicht standardisierten Datenaustausch zwischen Finanzdienstleistern und anderen Ecosystemen.
    
    **Implementierte Module (Datenbausteine):**
    - Initialisierung (Step 1) - Service Discovery & Cookie Consent
    - Produktauswahl (Step 2) - Bedürfnisbefriedigung & Konfiguration  
    - Selbstdeklaration (Step 3) - FATCA/MIFID II Information
    - Basisdaten (Step 4) - Identität, Kontakt, Adresse
    - Erweiterte Daten (Step 5) - Risiko-/Potenzialermittlung
    - Identifikation (Step 6) - E-ID kompatibel, VideoIdent
    - Background Checks (Step 7) - KYC, AML, PEP, Sanctions
    - Vertragsabschluss (Step 8) - AGB-Akzeptanz
    - Signatur-Services (Step 9) - QES/AES/eSignatur
    - Metadaten-Management (Step 10) - Verarbeitung & Integration
    
    **Use Cases Implementiert:**
    - UC1: Kundenbeziehungseröffnung (Banking Account Opening)
    - UC2: Re-Identifikation (Cross-Provider Identity Reuse)
    - UC3: Altersverifikation (Privacy-preserving Age Verification)
    - UC4: CLM von EVV-Endkunden (External Wealth Management Lifecycle)
    
    **Sicherheitsstandards:**
...
    
...
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - mTLS: []
paths:
...
  /authorize:
    get:
      summary: OAuth 2.1 Authorization Endpoint
      description: |
        FAPI 2.0 compliant OAuth 2.1 authorization endpoint. Supports PAR (Pushed Authorization Requests)
        and requires PKCE for all authorization flows.
      operationId: authorize
      tags:
        - OAuth 2.1 / OIDC
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
          description: Must be 'code' for authorization code flow
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: OAuth client identifier
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: Client redirect URI
        - name: scope
          in: query
          required: true
          schema:
            type: string
          description: Requested OAuth scopes
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: Client state parameter for CSRF protection
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
          description: PKCE code challenge (S256)
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum: [S256]
          description: PKCE code challenge method
        - name: nonce
          in: query
          schema:
            type: string
          description: OpenID Connect nonce
        - name: request_uri
          in: query
          schema:
            type: string
          description: PAR request URI (urn:ietf:params:oauth:request_uri:*)
      responses:
        '302':
          description: Redirect to redirect_uri with authorization code
        '400':
          description: Invalid authorization request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security: []
  /token:
    post:
      summary: OAuth 2.1 Token Endpoint
      description: |
        FAPI 2.0 compliant token endpoint supporting authorization_code and refresh_token grants.
        Requires mTLS or private_key_jwt client authentication.
      operationId: token
      tags:
        - OAuth 2.1 / OIDC
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - grant_type
                - client_id
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code, refresh_token]
                client_id:
                  type: string
                code:
                  type: string
                  description: Required for authorization_code grant
                redirect_uri:
                  type: string
                  format: uri
                  description: Required for authorization_code grant
                code_verifier:
                  type: string
                  description: PKCE code verifier
                refresh_token:
                  type: string
                  description: Required for refresh_token grant
                client_assertion_type:
                  type: string
                  enum: [urn:ietf:params:oauth:client-assertion-type:jwt-bearer]
                  description: For private_key_jwt authentication
                client_assertion:
                  type: string
                  description: JWT assertion for private_key_jwt authentication
            examples:
              authorization_code:
                summary: Authorization Code Grant
                value:
                  grant_type: authorization_code
                  client_id: bank-client-001
                  code: auth_code_123
                  redirect_uri: https://client.example.com/callback
                  code_verifier: pkce_verifier_123
      responses:
        '200':
          description: Access token response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid token request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - mTLS: []
        - private_key_jwt: []
...
  /consent:
    post:
      summary: Erstelle Consent-Request
      description: Initiiert einen Consent-Flow für Datenaustausch zwischen Institutionen
      operationId: createConsent
      tags:
        - Consent Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentRequest'
            examples:
              account_opening:
                summary: Kontoeröffnung
                value:
                  customerId: "sha256:a1b2c3d4e5f6..."
                  requestingInstitution: "CH-BANK-001"
                  providingInstitution: "CH-BANK-002"
                  dataCategories: ["basicData", "identification", "kycData"]
                  purpose: "accountOpening"
                  expiryDate: "2024-12-31T23:59:59Z"
                  customerContactMethod: "email"
      responses:
        '201':
          description: Consent erfolgreich erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /customer/check:
    post:
      summary: Kunden-Existenz prüfen
      description: Prüft ob ein Kunde bereits bei einer Institution identifiziert wurde (MVP Identifikation)
      operationId: checkCustomer
      tags:
        - Customer Data
      security:
        - mTLS: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCheckRequest'
            examples:
              basic_check:
                summary: Basis-Kundenprüfung
                value:
                  sharedCustomerHash: "sha256:abc123def456..."
                  basicData:
                    lastName: "Müller"
                    givenName: "Hans"
                    birthDate: "1985-03-15"
                    nationality: ["CH"]
      responses:
        '200':
          description: Kunde erfolgreich geprüft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerCheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /customer/data:
    post:
      summary: Kundendaten mit Consent abrufen
      description: Ruft spezifische Kundendatenmodule basierend auf gewährtem Consent ab
      operationId: getCustomerData
      tags:
        - Customer Data
      security:
        - mTLS: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerDataRequest'
      responses:
        '200':
          description: Kundendaten erfolgreich abgerufen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'

  # PORTFOLIO & EVV SERVICES
...
  /par:
    post:
      summary: Pushed Authorization Request
      description: Submits authorization request parameters to the authorization server (FAPI 2.0 required)
      operationId: createPushedAuthorizationRequest
      tags:
        - FAPI 2.0 OAuth
      security:
        - mTLS: []
        - PrivateKeyJWT: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PARRequest'
            examples:
              consent_request:
                summary: Consent Authorization Request
                value:
                  client_id: "CH-BANK-001"
                  response_type: "code"
                  scope: "consent:read consent:write customer:read"
                  redirect_uri: "https://client.example.com/callback"
                  state: "xyz"
                  code_challenge: "E9Melhoa2OwvFrEMTJguCHaoeK1t8URWbuGJSstw-cM"
                  code_challenge_method: "S256"
      responses:
        '201':
          description: PAR erfolgreich erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PARResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # HEALTH CHECK
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-Token mit Consent-Claims für API-Zugriff (FAPI 2.0 compliant)
    DPoP:
      type: http
      scheme: DPoP
      description: DPoP (Demonstrating Proof-of-Possession) für sender-constrained tokens
    mTLS:
      type: mutualTLS
      description: Mutual TLS für höchste Sicherheitsanforderungen (FAPI 2.0 compliant)
    private_key_jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Private Key JWT client authentication (RFC 7523) - alternative zu mTLS
    RegistrationToken:
      type: http
      scheme: bearer
      description: Registration access token für dynamic client registration
  responses:
    BadRequest:
      description: Ungültige Anfrage
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "INVALID_REQUEST"
            message: "Missing required field: customerId"
            timestamp: "2024-01-15T10:30:00Z"
    Unauthorized:
      description: Nicht autorisiert
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "UNAUTHORIZED"
            message: "Invalid or missing authentication token"
            timestamp: "2024-01-15T10:30:00Z"
    Forbidden:
      description: Zugriff verweigert
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "FORBIDDEN"
            message: "Insufficient permissions for requested operation"
            timestamp: "2024-01-15T10:30:00Z"
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Fehlercode
        message:
          type: string
          description: Fehlerbeschreibung
        timestamp:
          type: string
          format: date-time
          description: Zeitstempel des Fehlers
        details:
          type: object
          description: Zusätzliche Fehlerdetails
...
    ConsentRequest:
      type: object
      required:
        - customerId
        - requestingInstitution
        - dataCategories
        - purpose
        - expiryDate
      properties:
        customerId:
          type: string
          description: Eindeutige Kunden-ID (sharedCustomerHash)
          example: "sha256:abc123..."
        requestingInstitution:
          type: string
          description: Institution die Daten anfordert
          example: "CH-BANK-001"
        providingInstitution:
          type: string
          description: Institution die Daten bereitstellt
          example: "CH-BANK-002"
        dataCategories:
          type: array
          items:
            $ref: '#/components/schemas/DataCategory'
          description: Angeforderte Datenkategorien
        purpose:
          type: string
          enum: [accountOpening, creditAssessment, compliance, customerUpdate]
          description: Zweck der Datenverwendung
        expiryDate:
          type: string
          format: date-time
          description: Ablaufzeitpunkt des Consents
        customerContactMethod:
          type: string
          enum: [email, sms, app, postal]
          description: Bevorzugter Kontaktweg für Consent-Bestätigung
    ConsentResponse:
      type: object
      properties:
        consentId:
          type: string
          format: uuid
          description: Eindeutige Consent-ID
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        consentUrl:
          type: string
          format: uri
          description: URL für Kunden-Consent-Interface
        qrCode:
          type: string
          description: Base64-kodierter QR-Code für mobilen Consent
        expiryDate:
          type: string
          format: date-time
...
    DataCategory:
      type: string
      enum: 
        - basicData
        - identification
        - contactInformation
        - addressData
        - kycData
        - riskProfile
        - complianceData
        - extendedData

    # CUSTOMER DATA SCHEMAS
    CustomerCheckRequest:
      type: object
      required:
        - sharedCustomerHash
        - basicData
      properties:
        sharedCustomerHash:
          type: string
          description: SHA-256 Hash der Grunddaten für Matching
          example: "sha256:a1b2c3d4..."
        basicData:
          $ref: '#/components/schemas/BasicCustomerData'
    CustomerCheckResponse:
      type: object
      properties:
        match:
          type: boolean
          description: Ob ein Kunde gefunden wurde
        identificationDate:
          type: string
          format: date-time
          description: Datum der letzten Identifikation
        levelOfAssurance:
          type: string
          enum: [low, substantial, high]
          description: Sicherheitsniveau der Identifikation
        validUntil:
          type: string
          format: date-time
          description: Gültigkeit der Identifikation
...
    BasicCustomerData:
      type: object
      required:
        - lastName
        - givenName
        - birthDate
        - nationality
      properties:
        lastName:
          type: string
          description: Nachname
          example: "Müller"
          maxLength: 100
        givenName:
          type: string
          description: Vorname(n)
          example: "Hans Peter"
          maxLength: 100
        middleName:
          type: string
          description: Weitere Vornamen
          maxLength: 100
        title:
          type: string
          description: Titel
          example: "Dr."
          maxLength: 50
        birthDate:
          type: string
          format: date
          description: Geburtsdatum
          example: "1985-03-15"
        birthPlace:
          type: string
          description: Geburtsort
          example: "Zürich"
          maxLength: 100
        nationality:
          type: array
          items:
            type: string
            pattern: '^[A-Z]{2}$'
          description: Nationalität(en) (ISO 3166-1)
          example: ["CH", "DE"]
          minItems: 1
          maxItems: 5
        gender:
          type: string
          enum: [male, female, other, unknown]
        maritalStatus:
          type: string
          enum: [single, married, divorced, widowed, registered_partnership]
        language:
          type: string
          description: Bevorzugte Sprache (ISO 639-1)
          example: "de"
          pattern: '^[a-z]{2}$'
...
    PARRequest:
      type: object
      required:
        - client_id
        - response_type
        - redirect_uri
        - code_challenge
        - code_challenge_method
      properties:
        client_id:
          type: string
          description: Client Identifier
          maxLength: 255
        response_type:
          type: string
          enum: [code]
          description: OAuth 2.1 response type (only code allowed)
        scope:
          type: string
          description: OAuth scopes
          example: "consent:read consent:write customer:read"
        redirect_uri:
          type: string
          format: uri
          description: Client redirect URI
        state:
          type: string
          description: OAuth state parameter
          maxLength: 255
        code_challenge:
          type: string
          description: PKCE code challenge (FAPI 2.0 required)
          maxLength: 128
        code_challenge_method:
          type: string
          enum: [S256]
          description: PKCE code challenge method (S256 required)
        nonce:
          type: string
          description: OpenID Connect nonce
          maxLength: 255
    PARResponse:
      type: object
      required:
        - request_uri
        - expires_in
      properties:
        request_uri:
          type: string
          format: uri
          description: Request URI to use in authorization request
          example: "urn:ietf:params:oauth:request_uri:6esc_11ACC5bwc014ltc14eY22c"
        expires_in:
          type: integer
          description: Request URI lifetime in seconds (max 60 for FAPI 2.0)
          maximum: 60
          example: 60

    # REFERENZPROZESS SCHEMAS
...
    CustomerDataRequest:
      type: object
      required:
        - sharedCustomerHash
        - requestedModules
        - consentToken
      properties:
        sharedCustomerHash:
          type: string
          description: SHA-256 Hash des Kunden
        requestedModules:
          type: array
          items:
            type: string
            enum: [basisdaten_module, erweiterte_daten_module, identifikation_module, background_checks_module]
          description: Angeforderte Datenbausteine
        consentToken:
          type: string
          description: JWT-Token mit Consent-Nachweis
        purpose:
          type: string
          enum: [kundenbeziehungseroffnung, re_identification, compliance_update]
    CustomerDataResponse:
      type: object
      properties:
        availableModules:
          type: array
          items:
            type: string
        moduleData:
          type: object
          description: Bereitgestellte Datenbausteine
        verificationStatus:
          type: object
        lastUpdated:
          type: string
          format: date-time
...
    OAuthError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - server_error
            - temporarily_unavailable
        error_description:
          type: string
          description: Human-readable error description
        error_uri:
          type: string
          format: uri
          description: URI to error documentation
        timestamp:
          type: string
          format: date-time
    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: Refresh token for getting new access tokens
        token_type:
          type: string
          enum: [Bearer, DPoP]
          description: Token type (Bearer or DPoP)
        expires_in:
          type: integer
          description: Token lifetime in seconds
        scope:
          type: string
          description: Granted OAuth scopes
        id_token:
          type: string
          description: OpenID Connect ID token (JWT)
...